fix: use `StreamStart.CURRENT_TEST` for imperatively started processes
A misleading sanity check was failing when it was tried to retrieve one
of the process' streams with `StreamStart.CURRENT_TEST` when the process
was started manually / using the API during that very test (so if the
process wasn't already running before the start of the test).

This has been fixed and is now possible. In this case, the whole stream
is returned since the entirety of it took place during the current test.